gce/requirements.txt
python-dotenv
pillow
protobuf

torch
datasets
transformers
accelerate
bitsandbytes
sentencepiece
peft
trl

tensorboard
torchvision

evaluate
